Move search service into session module.
