+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/room/RoomModule.kt
@@ -25,7 +25,6 @@ import org.commonmark.renderer.html.HtmlRenderer
 import org.matrix.android.sdk.api.session.file.FileService
 import org.matrix.android.sdk.api.session.room.RoomDirectoryService
 import org.matrix.android.sdk.api.session.room.RoomService
-import org.matrix.android.sdk.api.session.room.search.SearchService
 import org.matrix.android.sdk.internal.session.DefaultFileService
 import org.matrix.android.sdk.internal.session.SessionScope
 import org.matrix.android.sdk.internal.session.room.alias.AddRoomAliasTask
@@ -62,9 +61,8 @@ import org.matrix.android.sdk.internal.session.room.relation.FindReactionEventFo
 import org.matrix.android.sdk.internal.session.room.relation.UpdateQuickReactionTask
 import org.matrix.android.sdk.internal.session.room.reporting.DefaultReportContentTask
 import org.matrix.android.sdk.internal.session.room.reporting.ReportContentTask
-import org.matrix.android.sdk.internal.session.room.search.DefaultSearchService
-import org.matrix.android.sdk.internal.session.room.search.DefaultSearchTask
-import org.matrix.android.sdk.internal.session.room.search.SearchTask
+import org.matrix.android.sdk.internal.session.search.DefaultSearchTask
+import org.matrix.android.sdk.internal.session.search.SearchTask
 import org.matrix.android.sdk.internal.session.room.state.DefaultSendStateTask
 import org.matrix.android.sdk.internal.session.room.state.SendStateTask
 import org.matrix.android.sdk.internal.session.room.tags.AddTagToRoomTask
@@ -124,9 +122,6 @@ internal abstract class RoomModule {
 
     @Binds
     abstract fun bindFileService(service: DefaultFileService): FileService
-
-    @Binds
-    abstract fun bindSearchService(service: DefaultSearchService): SearchService
 
     @Binds
     abstract fun bindCreateRoomTask(task: DefaultCreateRoomTask): CreateRoomTask
