+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/search/request/SearchRequestBody.kt
@@ -15,18 +15,13 @@
  * limitations under the License.
  */
 
-package org.matrix.android.sdk.internal.session.room.search.response
+package org.matrix.android.sdk.internal.session.search.request
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
-import org.matrix.android.sdk.api.session.events.model.Event
 
 @JsonClass(generateAdapter = true)
-data class SearchResponseItem(
-        // A number that describes how closely this result matches the search. Higher is closer.
-        @Json(name = "rank")
-        val rank: Int? = null,
-        // The event that matched.
-        @Json(name = "result")
-        val event: Event
+internal data class SearchRequestBody(
+        @Json(name = "search_categories")
+        val searchCategories: SearchRequestCategories
 )
