+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/search/request/SearchRequestCategories.kt
@@ -15,13 +15,13 @@
  * limitations under the License.
  */
 
-package org.matrix.android.sdk.internal.session.room.search.request
+package org.matrix.android.sdk.internal.session.search.request
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class SearchRequestCategories(
+internal data class SearchRequestCategories(
         // Mapping of category name to search criteria.
         @Json(name = "room_events")
         val roomEvents: SearchRequestRoomEvents? = null
