+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/search/request/SearchRequestOrder.kt
@@ -15,20 +15,15 @@
  * limitations under the License.
  */
 
-package org.matrix.android.sdk.internal.session.room.search.request
+package org.matrix.android.sdk.internal.session.search.request
 
-import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
-@JsonClass(generateAdapter = true)
-data class SearchRequestEventContext(
-        // How many events before the result are returned.
-        @Json(name = "before_limit")
-        val beforeLimit: Int? = null,
-        // How many events after the result are returned.
-        @Json(name = "after_limit")
-        val afterLimit: Int? = null,
-        // Requests that the server returns the historic profile information
-        @Json(name = "include_profile")
-        val includeProfile: Boolean? = null
-)
+/**
+ * Represents the order in which to search for results.
+ */
+@JsonClass(generateAdapter = false)
+internal enum class SearchRequestOrder(val value: String) {
+    RANK("rank"),
+    RECENT("recent")
+}
